+++ NEWS.md
@@ -1,4 +1,4 @@
-# tern 0.7.8.9022
+# tern 0.7.8.9024
 
 ### Enhancements 
 * Added `DescTools` `BinomDiffCI` function within `tern`.
@@ -66,6 +66,8 @@
 * Removed `tern:::` prefix and added `dontrun` to internal function examples.
 * Enhanced `s_coxph_pairwise` with generating log-rank p value by original 
   log-rank test, instead of Cox Proportional-Hazards Model.
+* Fixed bug in `s_ancova`, avoiding error when the first level of the arm factor 
+  is not the control arm.
 
 ### Documentation and NAMESPACE polishing
 *  Added stable badge for: